# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for managing tasks and sandboxes.",
        "title": "Platform Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/oidc/callback": {
            "get": {
                "description": "Verifies state and id_token returned by the OIDC provider, issues an app JWT, and redirects to the frontend with the token in the URL fragment.",
                "tags": [
                    "auth"
                ],
                "summary": "OIDC browser callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State JWT issued during /api/auth/oidc/login",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code from the OIDC provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "invalid state",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "id_token verification failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "code exchange failed or missing id_token",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/oidc/cli-callback": {
            "get": {
                "description": "Verifies the OIDC response for a CLI flow and stores the issued app token in Redis under the session ID.",
                "tags": [
                    "auth"
                ],
                "summary": "OIDC CLI callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "State JWT issued during cli-login",
                        "name": "state",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code from the OIDC provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentication successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid state or missing cli session",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "id_token verification failed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "code exchange failed or failed to store session",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/auth/oidc/cli-login": {
            "post": {
                "description": "Creates a pending Redis session keyed by session_id and returns the browser auth URL the CLI should open.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Start CLI OIDC login",
                "parameters": [
                    {
                        "description": "CLI session",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.cliLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.cliLoginResponse"
                        }
                    },
                    "400": {
                        "description": "session_id required",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.errorResponse"
                        }
                    },
                    "500": {
                        "description": "failed to build state or auth URL",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/oidc/cli-poll": {
            "get": {
                "description": "Returns the status of a CLI login session. When status is \"completed\" the response also contains the issued access token, and the session is deleted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Poll CLI OIDC session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CLI session ID",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.cliPollResponse"
                        }
                    },
                    "400": {
                        "description": "session_id required",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.errorResponse"
                        }
                    },
                    "404": {
                        "description": "session not found",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.errorResponse"
                        }
                    },
                    "500": {
                        "description": "redis error or invalid session data",
                        "schema": {
                            "$ref": "#/definitions/internal_oidc.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/oidc/login": {
            "get": {
                "description": "Redirects the browser to the OIDC provider's authorization endpoint. The optional ?redirect= is preserved through the flow.",
                "tags": [
                    "auth"
                ],
                "summary": "OIDC browser login",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Frontend redirect target after callback",
                        "name": "redirect",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "500": {
                        "description": "failed to build state or auth URL",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_oidc.cliLoginRequest": {
            "type": "object",
            "required": [
                "session_id"
            ],
            "properties": {
                "session_id": {
                    "type": "string"
                }
            }
        },
        "internal_oidc.cliLoginResponse": {
            "type": "object",
            "properties": {
                "auth_url": {
                    "type": "string"
                }
            }
        },
        "internal_oidc.cliPollResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "internal_oidc.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}